# src/management-system/src/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PROCEED Management System API",
    "description": "### About us\nPROCEED is a decentralized Business Process Management System (dBPMS) specialized on the creation, execution and monitoring of individual IoT processes. The execution is done decentralized without a controlling instance. Instead there are multiple machines that cooperate with each other to achieve the designed process goal. Thereby, all machines communicate their capabilities to the vicinity and the process is taking the best fitting one for its next execution ##step.\n\n### More resources: \n- [Documentation](https://docs.proceed-labs.org/)\n- [Wiki](https://github.com/PROCEED-Labs/proceed/wiki)",
    "contact": {
      "email": "proceed@snet.tu-berlin.de",
      "url": "mailto:proceed@snet.tu-berlin.de",
      "name": "proceed"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/PROCEED-Labs/proceed/blob/main/LICENSE"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://localhost:33083/api",
      "description": "Managent system in development"
    },
    {
      "url": "https://proceed.snet.tu-berlin.de/api",
      "description": ""
    }
  ],
...
  "security": [
    {
      "id": [],
      "x-csrf-token": [],
      "x-csrf": []
    },
    {}
  ],
  "paths": {
    "/process": {
      "get": {
        "operationId": "getProcess",
        "tags": ["Process"],
        "parameters": [
          {
            "in": "query",
            "name": "noBpmn",
            "description": "If set to true, no BPM XML is sent back",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of processes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/processResponse"
                  }
                },
                "examples": {
                  "Example 1": {
                    "value": [
                      {
                        "type": "process",
                        "description": "",
                        "owner": "id",
                        "processIds": ["Process_1muq5gw"],
                        "variables": [],
                        "departments": [],
                        "inEditingBy": [
                          {
                            "id": "SYQoPJlzlV7UMwlkAAA9",
                            "task": "null"
                          }
                        ],
                        "createdOn": "Fri, 07 Jul 2023 17:34:46 GMT",
                        "lastEdited": "Fri, 07 Jul 2023 17:36:10 GMT",
                        "shared": true,
                        "versions": [],
                        "definitionId": "_3fb07ee0-9f40-4f3c-91df-860376f9ee11",
                        "definitionName": "Example 1",
                        "bpmn": ""
                      },
                      {
                        "type": "process",
                        "description": "This is an example",
                        "owner": null,
                        "processIds": ["Process_1rlrc4d"],
                        "variables": [
                          {
                            "name": "varName",
                            "type": "object"
                          }
                        ],
                        "departments": [],
                        "inEditingBy": [],
                        "createdOn": "Fri, 05 Jun 2023 12:35:07 GMT",
                        "lastEdited": "Fri, 05 Jun 2023 12:36:09 GMT",
                        "shared": true,
                        "versions": [],
                        "definitionId": "_39d26168-0dee-459f-9d72-066c0aea0e8b",
                        "definitionName": "Example 2",
                        "bpmn": ""
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "500": {
            "description": "Failed to get processes"
          }
        },
        "description": "Get all processes you can view."
      },
      "post": {
        "operationId": "postProcess",
        "tags": ["Process"],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/processData"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "bpmn": {
                        "$ref": "#/components/schemas/bpmn"
                      }
                    }
                  }
                ],
                "required": ["departments", "bpmn"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Process created succesfuly",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/process",
                  "description": "Created Process"
                }
              }
            }
          },
          "303": {
            "description": "There already exists a process with the given id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/process",
                  "description": "Already existing process"
                }
              }
            }
          },
          "400": {
            "description": "The body of the request does not conform to the schma of a process"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "500": {
            "description": "Failed to create the process due an internal error"
          }
        },
        "description": "Post a process."
      }
    },
    "/process/{definitionId}": {
      "get": {
        "description": "Get a porcess by it's id.",
        "operationId": "getProcessById",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/processResponse"
                    }
                  ],
                  "required": ["bpmn"]
                }
              }
            }
          },
          "400": {
            "description": "Error getting the bpmn of the process"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "404": {
            "description": "No Process with the given definitionId was found"
          }
        }
      },
      "put": {
        "description": "Updates a **partial** set of properties (this means, that arbitrarily many properties can be left out). ",
        "operationId": "updateProcessById",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "description": {
                        "type": "string",
                        "description": "The description (the content of documentation element of the first process element in the process model)"
                      },
                      "definitionName": {
                        "type": "string",
                        "description": "The name as it is in the definitions element of the process"
                      }
                    }
                  },
                  {
                    "$ref": "#/components/schemas/processData"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "bpmn": {
                        "$ref": "#/components/schemas/bpmn"
                      }
                    }
                  }
                ],
                "description": "Partial set of key:value properties of the Process scheme"
              },
              "examples": {
                "example1": {
                  "value": {
                    "description": "New description"
                  }
                }
              }
            }
          },
          "description": ""
        },
        "responses": {
          "200": {
            "description": "Process updated succesfuly",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/process",
                  "description": "Updated Process"
                }
              }
            }
          },
          "400": {
            "description": "Request body doesn't contain a partial set of key:value properties of the Process scheme"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "404": {
            "description": "No Process with the given definitionId was found"
          }
        }
      },
      "delete": {
        "operationId": "deleteProcessById",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Process deleted succesfuly"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "404": {
            "description": "No Process with the given definitionId was found"
          }
        },
        "description": "Delete a process."
      }
    },
    "/process/{definitionId}/versions": {
      "get": {
        "operationId": "getProcessVersionsById",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/processVersion"
                  },
                  "description": "Array of the versions of the process"
                }
              }
            },
            "description": "Returns versions of a process"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "404": {
            "description": "No Process with the given definitionId was found"
          }
        },
        "description": "Get all the versions of a process."
      },
      "post": {
        "operationId": "postProcessVersionById",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bpmn": {
                    "type": "string",
                    "description": "BPMN XML"
                  }
                },
                "required": ["bpmn"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "description": "Post a new version of a process."
      }
    },
    "/process/{definitionId}/versions/{version}": {
      "get": {
        "operationId": "getVersionById",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "name": "version",
            "in": "path",
            "required": true,
            "description": "version number"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Returns a process"
          },
          "400": {
            "description": "{definitionId} or {version} are wrong"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "description": "Get a specific version of a process."
      }
    },
    "/process/{definitionId}/images": {
      "get": {
        "operationId": "getImageOfProcessById",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/png image/svg+xml image/jpeg": {
                "schema": {
                  "type": "string",
                  "description": "Requested image\n"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "description": "Get all images used in a process."
      }
    },
    "/process/{definitionId}/images/{imageFileName}": {
      "get": {
        "description": "Get a specific image og a process.",
        "operationId": "getImagesByProcessId",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "name": "imageFileName",
            "in": "path",
            "required": true,
            "description": "Filename of the image"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/image"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/process/{definitionId}/user-tasks": {
      "get": {
        "description": "Get all user tasks used in a process.",
        "operationId": "getUserTasksByProcessId",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "minItems": 0,
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        }
      }
    },
    "/process/{definitionId}/user-tasks/{userTaskFileName}": {
      "get": {
        "operationId": "getUserTaskByFilename",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "name": "userTaskFileName",
            "in": "path",
            "required": true,
            "description": "Filename of the user task html file"
          }
        ],
        "responses": {
          "200": {
            "description": "Plain html of the user task",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "description": "Get a specific user task of a process."
      },
      "put": {
        "description": "Update a specific user task of a process.\nIf userTaskFileName exists, then it is updated with the body of the request (200 response), if not, then the user task is created (201 response).",
        "operationId": "updateUserTaskByFilename",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "name": "userTaskFileName",
            "in": "path",
            "required": true,
            "description": "Filename of the user task html file"
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            },
            "text/html": {
              "schema": {
                "type": "string"
              }
            }
          },
          "description": "HTML for the userTaskFile in plain text"
        },
        "responses": {
          "200": {
            "description": "User task HTML updated"
          },
          "201": {
            "description": "User task HTML created in the server"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        }
      },
      "delete": {
        "operationId": "deleteUserTaskByFilename",
        "tags": ["Process"],
        "parameters": [
          {
            "$ref": "#/components/parameters/definitionId"
          },
          {
            "schema": {
              "type": "string"
            },
            "name": "userTaskFileName",
            "in": "path",
            "required": true,
            "description": "Filename of the user task html file"
          }
        ],
        "responses": {
          "200": {
            "description": "The request returns 200 wether the userTaskFile exists or not"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "description": "Delete a specific user task of a process."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "processData": {
        "title": "processData",
        "type": "object",
        "properties": {
          "departments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "variables": {
            "type": "array",
            "description": "The variables supposed to be used in the process",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "type": {
                  "enum": ["number", "string", "boolean", "array", "object", null]
                }
              },
              "required": ["name", "type"]
            }
          }
        }
      },
...
      "bpmn": {
        "type": "string",
        "description": "BPMN XML of the process\n"
      },
...
      "processResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/process"
          }
        ],
        "required": [
          "definitionId",
          "definitionName",
          "description",
          "departments",
          "variables",
          "createdOn",
          "lastEdited",
          "processIds",
          "type",
          "versions"
        ]
      },
...
    },
    "securitySchemes": {
      "x-csrf-token": {
        "name": "Cross-site request forgery token",
        "type": "apiKey",
        "in": "header",
        "description": "Cross-site request forgery token obtained after logging in."
      },
      "x-csrf": {
        "name": "Cross-site request forgery enabled",
        "type": "apiKey",
        "in": "header",
        "description": "Needs to be set to '1' in order for the x-csrf-token to be acknowledged by the server and make authenticated requests."
      },
      "id": {
        "name": "Session id",
        "type": "apiKey",
        "in": "cookie",
        "description": "Session id given by the server after logging in."
      }
    },
    "parameters": {
      "definitionId": {
        "name": "definitionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Id of an existing process"
      }
    },
    "responses": {
...
      "401_unauthenticated": {
        "description": "Authentication error.",
        "content": {
          "text/html": {
            "schema": {
              "type": "string",
              "description": "Authentication error."
            }
          }
        }
      },
      "403_validationFailed": {
        "description": "Request validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "x-examples": {
                "Example 1": {
                  "aditionalProperties": true
                }
              },
              "description": "Request validation error."
            }
          },
          "text/html": {
            "schema": {
              "type": "string",
              "description": "Authorization error."
            }
          }
        }
      }
    }
  }
}